+++ src/accounts/index.ts
@@ -26,6 +26,8 @@ export {
   type SignTypedDataReturnType,
   signTypedData,
 } from './utils/signTypedData.js'
+export type { SignTransactionArgs, SignTransactionReturnType } from './utils/signTransaction.js'
+export { sign } from './utils/sign.js'
 export { parseAccount } from './utils/parseAccount.js'
 export { publicKeyToAddress } from './utils/publicKeyToAddress.js'
 export { czech } from './wordlists/czech.js'
